Remove plotting functionality from ephem obj
And put into a dedicated plotter
